Transmission models extracted in the systematic review of articles related to Lassa fever.

This data set provides all extracted data for mathematical models applied to Lassa fever.

Usage

lassa_models

Format

An object of class spec_tbl_df (inherits from tbl_df, tbl, data.frame) with 52 rows and 12 columns.

Source

Doohan P, Jorgensen D, Naidoo TM, McCain K, Hicks JT, McCabe R, Bhatia S, Charniga K, Cuomo-Dannenburg G, Hamlet A, Nash RK, Nikitin D, Rawson T, Sheppard RJ, Unwin HJT, van Elsland S, Cori A, Morgenstern C, Imai-Eaton N; Pathogen Epidemiology Review Group. Lassa fever outbreaks, mathematical models, and disease parameters: a systematic review and meta-analysis. Lancet Glob Health. 2024 Dec;12(12):e1962-e1972. doi: 10.1016/S2214-109X(24)00379-6. PMID: 39577970.